+++ src/patient_data/tramp.hpp
@@ -16,12 +16,12 @@ public:
     ~Tramp_t();
     void setEnergies();
     void to_file(std::string file, std::string dir = std::string());
-    void scale_weights(std::vector<double> ratio);
+    void scale_weights(std::vector<float> ratio);
     void setWEPLs();
-    std::vector<double> getWEPLs();
+    std::vector<float> getWEPLs();
     void energy_to_internal();
-    void shift_energies(const std::vector<double>& e_, bool units = true);
-    void set_pos(const std::vector< Vector4_t<double> > p);
+    void shift_energies(const std::vector<float>& e_, bool units = true);
+    void set_pos(const std::vector< Vector4_t<float> > p);
     void print(unsigned int n);
     void print(unsigned int n0, unsigned int n1);
     void defaults();
@@ -32,13 +32,13 @@ public:
     std::vector<float> get_ypos();
 
     std::vector<Spot_t> spots;
-    std::vector<double> energies;
-    std::vector<double> energies_internal;
-    std::vector<double> wepls;
+    std::vector<float> energies;
+    std::vector<float> energies_internal;
+    std::vector<float> wepls;
 
     std::string machine;
 
-    double gigaprotons;
+    float gigaprotons;
     unsigned int nspots;
     std::string patient_id;
     std::string patient_first_name;
@@ -50,8 +50,8 @@ public:
     std::string gantry;
     std::string couch_rotation;
 
-    double z;
-    double zeff;
+    float z;
+    float zeff;
     
 private:
     bool internal_energy_set;
